layer mapping fixes.

+++ src/globals.h
@@ -37,7 +37,7 @@
  */
 /**@{  */
 
-typedef uint_fast32_t id_type; //!< id type is used for local mapping functions - there should be $n$ of them depending on CORE_SIZE
+typedef uint_fast64_t id_type; //!< id type is used for local mapping functions - there should be $n$ of them depending on CORE_SIZE
 typedef int32_t volt_type; //!< volt_type stores voltage values for membrane potential calculations
 typedef int64_t weight_type;//!< seperate type for synaptic weights.
 typedef uint32_t thresh_type;//!< Type for weights internal to the neurons.
